+++ CHANGELOG.md
@@ -1,18 +1,28 @@
 # yii2-grid
+===========
 
-## v1.1.1 [2015-11-15]
+v1.1.2 [2015-11-21]
+-------------------
+
+- Fix of panel HTML-template.
+
+v1.1.1 [2015-11-15]
+-------------------
 
 - Added ability sortering of visible a columns.
 
-## v1.1.0 [2015-11-13]
+v1.1.0 [2015-11-13]
+-------------------
 
 - Added ability realization custom storage (session, database and etc).
 - Rename manager class `GridSettings` to `Manager`.
 
-## v1.0.1 [2015-09-01]
+v1.0.1 [2015-09-01]
+-------------------
 
 - Added ability not add footer of grid column.
 
-## v1.0.0 [2015-07-24]
+v1.0.0 [2015-07-24]
+-------------------
 
 - First release.